moving seeded_lot to test_lots module
